Move non-manager files into data folder
